% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavesurfer.R
\name{annotator_app}
\alias{annotator_app}
\title{anotador_app}
\usage{
annotator_app(wavs_folder, annotations_folder = getwd(), labels = NULL)
}
\arguments{
\item{annotations_folder}{}
}
\description{
anotador_app
}
