% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_add_regions}
\alias{ws_add_regions}
\title{ws_add_regions}
\usage{
ws_add_regions(id, annotations = NULL, color = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{annotations}{data.frame.}
}
\description{
ws_add_regions
}
