% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_annotator}
\alias{ws_annotator}
\title{ws_annotator}
\usage{
ws_annotator(id, labels = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}
}
\description{
ws_annotator
}
