% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_cancel_ajax}
\alias{ws_cancel_ajax}
\title{ws_cancel_ajax}
\usage{
ws_cancel_ajax(id)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}
}
\description{
ws_cancel_ajax
}
