% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_clear_regions}
\alias{ws_clear_regions}
\title{ws_clear_regions}
\usage{
ws_clear_regions(id)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}
}
\description{
ws_clear_regions
}
