% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_destroy_minimap}
\alias{ws_destroy_minimap}
\title{ws_destroy_minimap}
\usage{
ws_destroy_minimap(id)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}
}
\description{
ws_destroy_minimap
}
