% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_load}
\alias{ws_load}
\title{ws_load}
\usage{
ws_load(id, url, peaks = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{url}{character of the url of the sound.}

\item{peaks}{numeric vector. Optional. the peaks to draw the soundwave.}
}
\description{
ws_load
}
