% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_microphone}
\alias{ws_microphone}
\title{ws_microphone}
\usage{
ws_microphone(id, parameters = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{parameters}{list.}
}
\description{
ws_microphone
}
