% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_microphone_stop}
\alias{ws_microphone_stop}
\title{ws_microphone_stop}
\usage{
ws_microphone_stop(id)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}
}
\description{
ws_microphone_stop
}
