% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_minimap}
\alias{ws_minimap}
\title{ws_minimap}
\usage{
ws_minimap(id, waveColor = "#999", progressColor = "#555", height = 30)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{waveColor}{a character. The color of the wave in hexadecimal representation. Default is '#999'.}

\item{progressColor}{a character. The color of wave behind the progress bar. In hexadecimal representation. Default is '#555'.}

\item{height}{a numeric. The height in pixels of the minimap.}
}
\description{
ws_minimap
}
