% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_on}
\alias{ws_on}
\title{ws_on}
\usage{
ws_on(id, event, callback = htmlwidgets::JS("function() {}"),
  replace = FALSE)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{event}{character. The name of the event.}

\item{callback}{JS() object. A JavaScript callback function.}

\item{replace}{logical. Should the current callback be replaced?}
}
\description{
ws_on
}
