% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_play}
\alias{ws_play}
\title{ws_play}
\usage{
ws_play(id, start = NULL, end = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{start}{numeric. The second of where it should start to play.}

\item{end}{numeric. The second of where it should stop to play.}
}
\description{
ws_play
}
