% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_region_labeller}
\alias{ws_region_labeller}
\title{ws_region_labeller}
\usage{
ws_region_labeller(id, labels, enable = TRUE)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{enable}{a logical. Default is TRUE.}
}
\description{
ws_region_labeller
}
