% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_seek_and_center}
\alias{ws_seek_and_center}
\title{ws_seek_and_center}
\usage{
ws_seek_and_center(id, progress)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{progress}{numeric. Seeks to a progress and centers view. [0..1] (0 = beginning, 1 = end).}
}
\description{
ws_seek_and_center
}
