% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_seek_to}
\alias{ws_seek_to}
\title{ws_seek_to}
\usage{
ws_seek_to(id, progress)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{progress}{numeric.}
}
\description{
ws_seek_to
}
