% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_set_height}
\alias{ws_set_height}
\title{ws_set_height}
\usage{
ws_set_height(id, height = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{height}{numeric. The height of the visualization.}
}
\description{
ws_set_height
}
