% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_set_mute}
\alias{ws_set_mute}
\title{ws_set_mute}
\usage{
ws_set_mute(id, mute = TRUE)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{mute}{logical. Mute or unmute the sound. Default is TRUE (mute).}
}
\description{
ws_set_mute
}
