% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_set_playback_rate}
\alias{ws_set_playback_rate}
\title{ws_set_playback_rate}
\usage{
ws_set_playback_rate(id, rate = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{rate}{numeric. The speed of playback. 0.5 is half speed, 1 is normal speed, 2 is double speed and so on.}
}
\description{
ws_set_playback_rate
}
