% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_set_progress_color}
\alias{ws_set_progress_color}
\title{ws_set_progress_color}
\usage{
ws_set_progress_color(id, color = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{color}{character. Hex (e.g. #4466fa) or name of a color (e.g. 'blue').}
}
\description{
ws_set_progress_color
}
