% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_set_volume}
\alias{ws_set_volume}
\title{ws_set_volume}
\usage{
ws_set_volume(id, new_volume = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{new_volume}{numeric. From 0 to 1.}
}
\description{
ws_set_volume
}
