% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_skip}
\alias{ws_skip}
\title{ws_skip}
\usage{
ws_skip(id, offset = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{offset}{numeric. How many seconds to skip. Can be nagative.}
}
\description{
ws_skip
}
