% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_skip_backward}
\alias{ws_skip_backward}
\title{ws_skip_backward}
\usage{
ws_skip_backward(id, seconds = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{seconds}{numeric. How many seconds to skip backward.}
}
\description{
ws_skip_backward
}
