% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_skip_forward}
\alias{ws_skip_forward}
\title{ws_skip_forward}
\usage{
ws_skip_forward(id, seconds = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{seconds}{numeric. How many seconds to skip forward.}
}
\description{
ws_skip_forward
}
