% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_timeline}
\alias{ws_timeline}
\title{ws_timeline}
\usage{
ws_timeline(id, parameters = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{parameters}{list.}
}
\description{
ws_timeline
}
