% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_toggle_mute}
\alias{ws_toggle_mute}
\title{ws_toggle_mute}
\usage{
ws_toggle_mute(id)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}
}
\description{
ws_toggle_mute
}
