% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_un}
\alias{ws_un}
\title{ws_un}
\usage{
ws_un(id, event)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{event}{character. The name of the event.}
}
\description{
ws_un
}
