% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_zoom}
\alias{ws_zoom}
\title{ws_zoom}
\usage{
ws_zoom(id, px_per_sec = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{px_per_sec}{numeric. Pixels per second.}
}
\description{
ws_zoom
}
