% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavesurfer.R
\name{runExample}
\alias{runExample}
\title{runExample}
\usage{
runExample(example = c("annotator", "microphone", "plugins",
  "decoration"), display.mode = "showcase")
}
\arguments{
\item{example}{The example app to launch: "annotator", "microphone", "plugins", "decoration".}

\item{display.mode}{The mode in which to display the example. Defaults to showcase, but may be set to normal to see the example without code or commentary.}
}
\description{
Launch shiny example applications.
}
