% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavesurfer.R
\name{wavesurfer}
\alias{wavesurfer}
\title{<Add Title>}
\usage{
wavesurfer(audio, audioContext = NULL, audioRate = 1,
  audioScriptProcessor = NULL, autoCenter = TRUE,
  backend = "WebAudio", backgroundColor = NULL, barHeight = 1,
  barGap = NULL, barWidth = NULL, closeAudioContext = FALSE,
  container = NULL, cursorColor = "#333", cursorWidth = 1,
  duration = NULL, fillParent = TRUE, forceDecode = FALSE,
  hideScrollbar = FALSE, interact = TRUE, loopSelection = TRUE,
  maxCanvasWidth = 4000, mediaControls = FALSE, mediaType = "audio",
  minPxPerSec = 20, normalize = FALSE, partialRender = FALSE,
  pixelRatio = NULL, plugins = NULL, progressColor = "#555",
  removeMediaElementOnDestroy = TRUE, renderer = NULL,
  responsive = FALSE, rtl = FALSE, scrollParent = FALSE,
  skipLength = 2, splitChannels = FALSE, waveColor = "#999",
  xhr = list(), width = NULL, height = NULL, elementId = NULL,
  annotations = NULL)
}
\description{
<Add Description>
}
