% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_add_regions}
\alias{ws_add_regions}
\title{ws_add_regions}
\usage{
ws_add_regions(id, annotations = NULL, color = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{annotations}{a data.frame with columns "sound_id" (character), "region_id"
(character), "start" (numeric), "end" (numeric), "label" (character). The rows
represents the annotated regions of the audio.}

\item{color}{a character with the following format: 'rgb(r, g, b, a)'. Default is 'rgb(250, 200, 10, 0.5)'.}
}
\description{
ws_add_regions
}
