% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_cursor}
\alias{ws_cursor}
\title{ws_cursor}
\usage{
ws_cursor(id, parameters = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{parameters}{list.}
}
\description{
ws_cursor
}
