% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_cursor}
\alias{ws_cursor}
\title{ws_cursor}
\usage{
ws_cursor(id, showTime = TRUE, opacity = 1,
  customShowTimeStyle = list(`background-color` = "#000", color = "#fff",
  padding = "2px", `font-size` = "10px"))
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{showTime}{a logical. If TRUE displays the time next to the cursor. Defaults to TRUE.}

\item{opacity}{a numeric. The opacity/transparency. Default is 1.}

\item{customShowTimeStyle}{a list with custom styles which are applied to the cursor time element, such
'background-color', color, padding, 'font-size', etc.}
}
\description{
ws_cursor
}
