% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_regions}
\alias{ws_regions}
\title{ws_regions}
\usage{
ws_regions(id, dragSelection = TRUE)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{dragSelection}{a lofical. If TRUE lets you create regions by selecting areas
of the waveform with mouse. Useful for annotations. Default is TRUE.}
}
\description{
ws_regions
}
