% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_set_volume}
\alias{ws_set_volume}
\title{ws_set_volume}
\usage{
ws_set_volume(id, new_volume = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{new_volume}{numeric. The volume of the playback. 0 is mute, 1 is normal volume, 2 is twice loud and so on.}
}
\description{
ws_set_volume
}
