% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ws_spectrogram}
\alias{ws_spectrogram}
\title{ws_spectrogram}
\usage{
ws_spectrogram(id, parameters = NULL)
}
\arguments{
\item{id}{wavesurfer object or a character of its respective element id.}

\item{parameters}{list.}
}
\description{
ws_spectrogram
}
